A algorithm for feature selection
Rough set theory deals with concepts that are vague and creates an approximate description of patterns for data processing.  It basically defines the crisp set with rough representation. The quick reduct starts with the complete initial set and removes one feature at a time by ensuring identical predictive capability of the decision feature as that of the original feature set.